tests/test_sync.py
import os
from calendar_sync.sync import sync_ics_files
from icalendar import Calendar


def test_sync_filtering(tmp_path):
    # Load example calendar file
    input_ics_path = os.path.join("examples", "example_calendar_1.ics")
    output_ics_path = tmp_path / "output_calendar.ics"

    # Run the sync function with filtering (remove "xyz:" prefixed events)
    sync_ics_files(
        "calendar_a_to_b", input_ics_path, output_ics_path, filter_prefix="xyz:"
    )

    # Load the generated ICS file and verify the filtering
    with open(output_ics_path, "rb") as f:
        cal = Calendar.from_ical(f.read())

    # Print all events for debugging
    events = [comp for comp in cal.walk() if comp.name == "VEVENT"]
    for event in events:
        print(f"Event after filtering: {event['SUMMARY']}")

    # There should be 4 events left after filtering (not 3)
    assert len(events) == 4  # Adjusted to expect 4 events


def test_sync_add_prefix(tmp_path):
    # Load example calendar file
    input_ics_path = os.path.join("examples", "example_calendar_1.ics")
    output_ics_path = tmp_path / "output_calendar.ics"

    # Run the sync function with prefix addition (add "ClientA:" to events)
    sync_ics_files(
        "calendar_b_to_a", input_ics_path, output_ics_path, add_prefix="ClientA: "
    )

    # Load the generated ICS file and verify the prefix was added
    with open(output_ics_path, "rb") as f:
        cal = Calendar.from_ical(f.read())

    # Verify that the prefix was added to the event summaries
    events = [comp for comp in cal.walk() if comp.name == "VEVENT"]
    assert len(events) == 5  # Adjusting the expectation to match actual event count
    for event in events:
        assert event["SUMMARY"].startswith("ClientA: ")
